# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Iremeapp REST APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/discipline/{school_id}": {
            "get": {
                "tags": [
                    "Get displine marks API"
                ],
                "operationId": "getConductMarks",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID",
                        "required": true
                    },
                    {
                        "name": "class_id",
                        "in": "query",
                        "description": "Class Id",
                        "required": true
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Displine marks are successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/discipline": {
            "post": {
                "tags": [
                    "Save discipline marks API"
                ],
                "operationId": "saveConductMarks",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "class_id",
                                    "limit",
                                    "page"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "type": "text"
                                    },
                                    "class_id": {
                                        "type": "number"
                                    },
                                    "limit": {
                                        "type": "number"
                                    },
                                    "page": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Displine marks are successfully imported",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/discipline/update": {
            "post": {
                "tags": [
                    "Update discipline marks API"
                ],
                "operationId": "updateConductMarks",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "class_id",
                                    "student_id"
                                ],
                                "properties": {
                                    "school_id": {
                                        "type": "text"
                                    },
                                    "class_id": {
                                        "type": "number"
                                    },
                                    "student_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Displine marks are successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/discipline/bulk_edit": {
            "post": {
                "tags": [
                    "Bulk update discipline API"
                ],
                "operationId": "updateBulkConductMarks",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "class_id",
                                    "student_id"
                                ],
                                "properties": {
                                    "school_id": {
                                        "type": "text"
                                    },
                                    "class_id": {
                                        "type": "number"
                                    },
                                    "student_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Displine marks are successfully updated in bulk",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000//oauth/authorize",
                        "tokenUrl": "http://localhost:8000//oauth/token",
                        "refreshUrl": "http://localhost:8000//token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}